feat: 산업 매핑 100% 완료 및 group_code 체계 정비
  - 회사 산업 매핑 100% 완료 (1,228개 전체 매핑)
  - group_code 체계 정비: 21개 주요 그룹 + 기타 3개 (AGRI, MINING, ETC)
  - HOLDING(지주회사), EVENT(이벤트/테마) 그룹 추가
  - ETF 섹터 클러스터 재생성 (161개 GROUP_CODE + 12개 ASSET_TYPE)

+++ db/init.sql
@@ -444,7 +444,7 @@ CREATE TABLE "etf_other_composition" (
     "identifier_type" VARCHAR(20),                -- ISIN / TICKER / CUSTOM
     "identifier_value" VARCHAR(30),               -- 식별값
     "weight" DECIMAL(6,3),
-    "market_value" INTEGER,
+    "market_value" BIGINT,
     "created_at" TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT "fk_other_composition_etf" FOREIGN KEY ("etf_id") REFERENCES "etf"("id") ON DELETE CASCADE
 );
